Add ft_strcmp

+++ ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: ajakob <ajakob@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/09 20:12:09 by ajakob            #+#    #+#             */
-/*   Updated: 2023/08/09 20:13:23 by ajakob           ###   ########.fr       */
+/*   Updated: 2023/08/09 20:15:22 by ajakob           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@ int	ft_strcmp(const char *s1, const char *s2)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
-	size_t			i;
+	int				i;
 
 	str1 = (unsigned char *)s1;
 	str2 = (unsigned char *)s2;
